refactor: clean up staking module naming, remove dead code, extract atom
