+++ packages/uikit/src/components/staking/StakingForm.tsx
@@ -6,7 +6,8 @@ import { styled } from 'styled-components';
 import { AppRoute } from '../../libs/routes';
 import { useAtom } from '../../libs/useAtom';
 import { useEncodeStakingDeposit } from '../../state/staking/useEncodeStaking';
-import { stakingSelectedPool$, useStakingPools } from '../../state/staking/useStakingPools';
+import { stakingSelectedPool$ } from '../../state/staking/stakingAtoms';
+import { useStakingPools } from '../../state/staking/useStakingPools';
 import { TonTransactionNotification } from '../connect/TonTransactionNotification';
 import { useNavigate } from '../../hooks/router/useNavigate';
 import { StakingAmountInput } from './StakingAmountInput';
