+++ packages/uikit/src/components/staking/UnstakeForm.tsx
@@ -14,7 +14,8 @@ import { formatter, formatFiatCurrency } from '../../hooks/balance';
 import { useRate, useFormatFiat } from '../../state/rates';
 import { useJettonBalance } from '../../state/jetton';
 import { useTonBalance } from '../../state/wallet';
-import { stakingSelectedPool$, useStakingPools } from '../../state/staking/useStakingPools';
+import { stakingSelectedPool$ } from '../../state/staking/stakingAtoms';
+import { useStakingPools } from '../../state/staking/useStakingPools';
 import { useEncodeStakingUnstake } from '../../state/staking/useEncodeStaking';
 import { useStakingCycleCountdown } from '../../state/staking/useStakingCycleCountdown';
 import { useStakingPosition } from '../../state/staking/useStakingPosition';
