+++ packages/uikit/src/state/staking/usePromotedStakingPool.ts
@@ -1,10 +1,9 @@
 import { useMemo } from 'react';
-import { PoolImplementationType, PoolInfo } from '@tonkeeper/core/dist/tonApiV2';
+import { PoolInfo } from '@tonkeeper/core/dist/tonApiV2';
 import { useStakingPools } from './useStakingPools';
+import { getStakingPoolProvider } from './poolBranding';
 
-const isPromotedPool = (pool: PoolInfo) =>
-    pool.implementation === PoolImplementationType.LiquidTf &&
-    pool.name.toLowerCase().includes('tonstakers');
+const isPromotedPool = (pool: PoolInfo) => getStakingPoolProvider(pool) === 'tonstakers';
 
 export const usePromotedStakingPool = (): PoolInfo | undefined => {
     const { data: pools } = useStakingPools();
